Add Data Validation to Stock and Comment DTOs and Controllers

+++ api/Controllers/CommentController.cs
@@ -20,25 +20,30 @@ namespace api.Controllers
         [HttpGet]
         public async Task<IActionResult> GetAll()
         {
+            if(!ModelState.IsValid) return BadRequest(ModelState);
+
             var comments = await _commentRepository.GetAllAsync();
             var commentDto = comments.Select(s => s.ToCommentDTO());
 
             return Ok(commentDto);
         }
 
-        [HttpGet("{id}")]
+        [HttpGet("{id:int}")]
         public async Task<IActionResult> GetById([FromRoute] int id)
         {
-            var comment = await _commentRepository.GetByIdAsync(id);
+            if (!ModelState.IsValid) return BadRequest(ModelState);
 
+            var comment = await _commentRepository.GetByIdAsync(id);
             if (comment == null) return NotFound();
 
             return Ok(comment.ToCommentDTO());
         }
 
-        [HttpPost("{stockId}")]
+        [HttpPost("{stockId:int}")]
         public async Task<IActionResult> Create([FromRoute] int stockId, CreateCommentDTO createCommentDTO)
         {
+            if (!ModelState.IsValid) return BadRequest(ModelState);
+
             if (!await _stockRepository.StockExists(stockId)) return BadRequest("Stock does not exist!");
 
             var commentModel = createCommentDTO.ToCommentFromCreate(stockId);
@@ -48,22 +53,24 @@ namespace api.Controllers
         }
 
         [HttpPut]
-        [Route("{id}")]
+        [Route("{id:int}")]
         public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateCommentRequestDTO updateCommentRequestDTO)
         {
-            var comment = await _commentRepository.UpdateAsync(id, updateCommentRequestDTO.ToCommentFromUpdate());
+            if (!ModelState.IsValid) return BadRequest(ModelState);
 
+            var comment = await _commentRepository.UpdateAsync(id, updateCommentRequestDTO.ToCommentFromUpdate());
             if (comment == null) return NotFound("Comment does not exist!");
 
             return Ok(comment.ToCommentDTO());
         }
 
         [HttpDelete]
-        [Route("{id}")]
+        [Route("{id:int}")]
         public async Task<IActionResult> Delete([FromRoute] int id)
         {
-            var commentModel = await _commentRepository.DeleteAsync(id);
+            if (!ModelState.IsValid) return BadRequest(ModelState);
 
+            var commentModel = await _commentRepository.DeleteAsync(id);
             if (commentModel == null) return NotFound("Comment does not exist!");
 
             return Ok(commentModel);
